Fixup code by testing against an actual streamlit app
